requirements.txt
transformers
sentence-transformers
torch
faiss-cpu
scikit-learn
rank-bm25
numpy
nltk
docx
PyPDF2
pdfplumber
pypandoc

